requirements.txt
tweepy==3.8.0
discord.py==1.3.3
firebase-admin==4.1.0
pyaml==19.12.0